Refactor: Refactor Header component
* Unify NavButton to Header

* Modify name of Header style file
