+++ src/components/Header/Header.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { NavButton } from '~/components/Header/NavButton/NavButton';
 import SVGIconBtn from '~/components/SVGIconBtn/SVGIconBtn';
 import * as colors from '~/styles/common/colors';
 import * as sizes from '~/styles/common/sizes';
-import { MainLogo, StyledHeader } from './StyledHeader';
+import { MainLogo, NavBtnWrapper, StyledHeader } from './Header.styles';
 import { LoginIcon, LogoutIcon } from '~/assets';
 
 interface IProps {
@@ -14,6 +13,18 @@ interface IProps {
     onLogout(): Promise<void>;
 }
 
+const NavButton = ({ isNavOpen, handleNav }: Pick<IProps, 'isNavOpen' | 'handleNav'>) => {
+    const lineLen = new Array(3).fill("");
+
+    return (
+        <div>
+            <NavBtnWrapper isOpen={isNavOpen} onClick={handleNav}>
+                {lineLen.map((_, i) => <span key={i} className='line'></span>)}
+            </NavBtnWrapper>
+        </div>
+    );
+}
+
 const Header = ({ isNavOpen, isLoggedIn, handleNav, handleLoginBox, onLogout }: IProps) => {
     const onClick = () => {
         isLoggedIn ? onLogout() : handleLoginBox();
@@ -21,12 +32,10 @@ const Header = ({ isNavOpen, isLoggedIn, handleNav, handleLoginBox, onLogout }: 
 
     return (
         <StyledHeader>
-            <div>
-                <NavButton 
-                    isNavOpen={isNavOpen} 
-                    handleNav={handleNav}
-                />
-            </div>
+            <NavButton
+                isNavOpen={isNavOpen} 
+                handleNav={handleNav}
+            />
             <div>
                 <MainLogo as="a">
                     <h1>myFavoriteQuotes</h1>
